stupid comiitdrf daad dadad aadd

+++ app/src/main/java/tw/com/lixin/wm_casino/dataModels/TableLogData.java
@@ -13,7 +13,7 @@ public class TableLogData {
         public int groupID = -99;
         public Boolean bOk;
         public float balance;
-        public Map<Integer, String> dtOdds;
+        public Map<Long, String> dtOdds;
         public int maxBet01;
         public int maxBet02;
         public int maxBet03;
